fix all tensor device alloc dealloc counters and prints
Changes the device allocation and deallocation to take in
real tensors which allow me to do a better job forcing
everything to notify on its allocation and deallocation
counter.

Should help figure out whats up with the memory leaks in
convolution forward and backward.

Also further changes toward unifying on in in the whole
project.

+++ src/loss_softmax.c
@@ -4,7 +4,7 @@
 #include <math.h>
 
 label_t *label_make_random(int nr_elem, int range){
-  label_t *labels = (label_t*)mem_alloc(sizeof(label_t)*nr_elem);
+  label_t *labels = (label_t*)mem_alloc(sizeof(label_t)*(uint)nr_elem);
   for(int i =0; i< nr_elem; i++){
     labels[i] = ((int)rand())%range;
   }
